+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1] - TODO: Add release date.
+# Changelog
 
-* TODO: Describe initial release.
+## 0.1.0 - 2019-01-03
+
+### Added
+- Proper example of how this package can be used.
+
+### Changed
+- Package description.